Move stepHasTag to trait
